ML_Framework/requirements.txt
# 基础科学计算包
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习包
scikit-learn>=1.0.0

# 可视化包
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 配置文件支持
pyyaml>=6.0

# 数据处理
xlrd>=2.0.0
openpyxl>=3.0.0

# 进度条
tqdm>=4.62.0

# 日志和调试
colorlog>=6.0.0

# 数值计算优化
numba>=0.56.0

# 超参数优化
optuna>=3.0.0
cma>=3.2.0

# 可选：Jupyter notebook支持
jupyter>=1.0.0
ipykernel>=6.0.0 